pyproject.toml
[project]
name = "sassy-factcheck-bot"
version = "0.1.0"
description = "Sassy fact-checking Instagram DM bot that roasts bad takes with citations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==2.8.1",
    "anthropic>=0.34.0",
    "openai>=1.40.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "aiofiles>=23.0.0",
    "asyncio-mqtt>=0.16.0",
    "websockets>=12.0",
    "Pillow>=10.0.0",
    "instagrapi>=2.1.5",
    "httpx>=0.25.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]